requirements.txt
oracledb
PyYAML
